// frontend/src/api/generated/serializers/Bootstrap.ts
// Typelizer digest b2edb63e340cf6e8793853d715b1f1f1
//
// DO NOT MODIFY: This file was automatically generated by Typelizer.
import type {Auth, AppConfig, Flags, Turnstile} from "./index"

type Bootstrap = {
  locale: string;
  locales: Array<string>;
  i18nVersion: string;
  auth: Auth | null;
  app: AppConfig;
  flags: Flags;
  turnstile: Turnstile;
}

export type { Bootstrap };